+++ CHANGELOG.md
@@ -42,6 +42,7 @@ This is used for recording Invader's changes. This changelog is based on
 - invader-build: Fixed some issues with scenery not spawning on MCC
 - invader-build: Fixed uncompressed size having the wrong percentage
 - invader-build: Fixed an issue with rain not appearing in some maps
+- invader-build: Fixed an issue with detail objects not appearing in some maps
 - invader-edit-qt: Fixed crashing if a tag directory contained files that could
   not be accessed in the filesystem
 - invader-edit-qt: Fixed 8-bit color being displayed in the wrong order